ledger: Prevent pager from being spawned
The SH module by default allocates a pseudo TTY. This causes ledger to
spawn a pager, causing shell escape codes to be returned in the output,
which then causes problems matching output.

This disables the output tty in the sh command, which will fix the error
showing up when trying to load the homepage.

See also: https://stackoverflow.com/a/20128598/20261